Add initial Latex support
Client converts the latex to markdown before sending.
This is vital so they can properly sign the content hash (unless we want
to store the original and conversion on the server? But then it's hard
to verify things)
NOTE: currently we render md to html on the server. However, we probably
want to do the rendering client side.